Catch up fork

+++ modules/apps/document-library/document-library-uad-test/src/testIntegration/java/com/liferay/document/library/uad/test/DLFileEntryUADTestUtil.java
@@ -16,6 +16,7 @@ package com.liferay.document.library.uad.test;
 
 import com.liferay.document.library.kernel.model.DLFileEntry;
 import com.liferay.document.library.kernel.model.DLFolder;
+import com.liferay.document.library.kernel.model.DLFolderConstants;
 import com.liferay.document.library.kernel.service.DLAppLocalService;
 import com.liferay.document.library.kernel.service.DLFileEntryLocalService;
 import com.liferay.document.library.kernel.service.DLFolderLocalService;
@@ -25,7 +26,6 @@ import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
 import com.liferay.portal.kernel.test.util.TestDataConstants;
-import com.liferay.portal.kernel.test.util.TestPropsValues;
 import com.liferay.portal.kernel.util.ContentTypes;
 
 import java.io.ByteArrayInputStream;
@@ -41,26 +41,38 @@ public class DLFileEntryUADTestUtil {
 	public static DLFileEntry addDLFileEntry(
 			DLAppLocalService dlAppLocalService,
 			DLFileEntryLocalService dlFileEntryLocalService,
-			DLFolderLocalService dlFolderLocalService, long userId)
+			DLFolderLocalService dlFolderLocalService, long userId,
+			long groupId)
+		throws Exception {
+
+		DLFolder dlFolder = dlFolderLocalService.addFolder(
+			userId, groupId, groupId, false, 0L, RandomTestUtil.randomString(),
+			RandomTestUtil.randomString(), false,
+			ServiceContextTestUtil.getServiceContext());
+
+		return addDLFileEntry(
+			dlAppLocalService, dlFileEntryLocalService, dlFolder.getFolderId(),
+			userId, groupId);
+	}
+
+	public static DLFileEntry addDLFileEntry(
+			DLAppLocalService dlAppLocalService,
+			DLFileEntryLocalService dlFileEntryLocalService, long dlFolderId,
+			long userId, long groupId)
 		throws Exception {
 
 		ServiceContext serviceContext =
 			ServiceContextTestUtil.getServiceContext();
-
-		DLFolder dlFolder = dlFolderLocalService.addFolder(
-			userId, TestPropsValues.getGroupId(), TestPropsValues.getGroupId(),
-			false, 0L, RandomTestUtil.randomString(),
-			RandomTestUtil.randomString(), false, serviceContext);
 
 		byte[] bytes = TestDataConstants.TEST_BYTE_ARRAY;
 
 		InputStream is = new ByteArrayInputStream(bytes);
 
 		FileEntry fileEntry = dlAppLocalService.addFileEntry(
-			userId, dlFolder.getRepositoryId(), dlFolder.getFolderId(),
-			RandomTestUtil.randomString(), ContentTypes.TEXT_PLAIN,
-			RandomTestUtil.randomString(), StringPool.BLANK, StringPool.BLANK,
-			is, bytes.length, serviceContext);
+			userId, groupId, dlFolderId, RandomTestUtil.randomString(),
+			ContentTypes.TEXT_PLAIN, RandomTestUtil.randomString(),
+			StringPool.BLANK, StringPool.BLANK, is, bytes.length,
+			serviceContext);
 
 		return dlFileEntryLocalService.getFileEntry(fileEntry.getFileEntryId());
 	}
@@ -73,13 +85,19 @@ public class DLFileEntryUADTestUtil {
 		throws Exception {
 
 		for (DLFileEntry dlFileEntry : dlFileEntries) {
-			if (dlFileEntryLocalService.fetchDLFileEntry(
-					dlFileEntry.getFileEntryId()) != null) {
+			DLFileEntry existingDLFileEntry =
+				dlFileEntryLocalService.fetchDLFileEntry(
+					dlFileEntry.getFileEntryId());
 
+			if (existingDLFileEntry != null) {
 				dlAppLocalService.deleteFileEntry(dlFileEntry.getFileEntryId());
 			}
 
-			dlFolderLocalService.deleteFolder(dlFileEntry.getFolderId());
+			long dlFolderId = dlFileEntry.getFolderId();
+
+			if (dlFolderId != DLFolderConstants.DEFAULT_PARENT_FOLDER_ID) {
+				dlFolderLocalService.deleteFolder(dlFileEntry.getFolderId());
+			}
 		}
 	}
