+++ modules/apps/subscription/subscription-service/src/main/java/com/liferay/subscription/internal/model/listener/SubscriptionLayoutModelListener.java
@@ -32,13 +32,17 @@ public class SubscriptionLayoutModelListener extends BaseModelListener<Layout> {
 
 	@Override
 	public void onAfterRemove(Layout layout) {
+		if (layout == null) {
+			return;
+		}
+
 		try {
 			_subscriptionLocalService.deleteSubscriptions(
 				layout.getCompanyId(), Layout.class.getName(),
 				layout.getPlid());
 		}
-		catch (PortalException pe) {
-			throw new ModelListenerException(pe);
+		catch (PortalException portalException) {
+			throw new ModelListenerException(portalException);
 		}
 	}
